+++ src/pieces/sortings/release-date.js
@@ -1,7 +1,7 @@
 const releaseDate = {
   label: 'release date',
-  defaultDirectionLabel: 'newest first',
-  reverseDirectionLabel: 'oldest first',
+  defaultDirectionLabel: 'newest',
+  reverseDirectionLabel: 'oldest',
   fn: pieces => pieces.slice(0).sort((a, b) => b.releaseDate - a.releaseDate),
 };